# Optimize Label parsing calls