--- a/core/src/main/java/jenkins/model/Jenkins.java
+++ b/core/src/main/java/jenkins/model/Jenkins.java
@@ -2115,6 +2115,16 @@ public class Jenkins extends AbstractCIBase implements DirectlyModifiableTopLeve
     }
 
     /**
+     * Returns the label atom of the given name, only if it already exists.
+     * @return non-null if the label atom already exists.
+     */
+    public @Nullable LabelAtom tryGetLabelAtom(@CheckForNull String name) {
+        if (name == null)  return null;
+        return (LabelAtom) labels.get(name);
+    }
+
+
+    /**
      * Gets all the active labels in the current system.
      */
     public Set<Label> getLabels() {