@MarkEWaite 2023-08-24
@Vlatombe thanks for this improvement.  It appears that the improvement causes one of the automated tests in the workflow durable task plugin to fail as shown in https://github.com/jenkinsci/bom/pull/2412/checks?check_run_id=16175702436 

Thanks to @daniel-beck for detecting the issue and mentioning it in https://github.com/jenkinsci/bom/pull/2412#issuecomment-1691512328

Can you investigate that failure and propose a fix for the tests in the workflow durable task plugin so that a new version of workflow durable task plugin can be included in the plugin bill of materials?  Without that fix, the plugin compatibility tests used for the plugin bill of materials will fail and block the next release of the plugin bill of materials.