@Vlatombe 2023-08-24
Hi @MarkEWaite this test failure has been addressed already - https://github.com/jenkinsci/workflow-durable-task-step-plugin/pull/335