@MarkEWaite 2023-08-24
Thanks very much @Vlatombe !  I've triggered a dependabot update of the plugin bill of materials to bring the new release of the plugin into the BOM.