@aaronmell 2023-08-30
FYI, this change likely broken our Jenkins instance.  Jenkins fails to start after upgrading, and we are getting a null reference error

```
Error
java.lang.NullPointerException
	at java.base/java.util.HashSet.<init>(HashSet.java:119)
	at hudson.model.Node.getAssignedLabels(Node.java:326)
	at jenkins.model.Jenkins.lambda$trimLabels$3(Jenkins.java:2323)
	at java.base/java.util.ArrayList.forEach(ArrayList.java:1541)
	at jenkins.model.Jenkins.trimLabels(Jenkins.java:2323)
	at jenkins.model.Jenkins.trimLabels(Jenkins.java:2303)
	at jenkins.model.Nodes$6.run(Nodes.java:352)
	at hudson.model.Queue._withLock(Queue.java:1397)
	at hudson.model.Queue.withLock(Queue.java:1271)
	at jenkins.model.Nodes.load(Nodes.java:346)
	at jenkins.model.Jenkins$12.run(Jenkins.java:3488)
	at org.jvnet.hudson.reactor.TaskGraphBuilder$TaskImpl.run(TaskGraphBuilder.java:177)
	at org.jvnet.hudson.reactor.Reactor.runTask(Reactor.java:305)
	at jenkins.model.Jenkins$5.runTask(Jenkins.java:1170)
	at org.jvnet.hudson.reactor.Reactor$2.run(Reactor.java:221)
	at org.jvnet.hudson.reactor.Reactor$Node.run(Reactor.java:120)
	at jenkins.security.ImpersonatingExecutorService$1.run(ImpersonatingExecutorService.java:68)
	at java.base/java.util.concurrent.ThreadPoolExecutor.runWorker(ThreadPoolExecutor.java:1128)
	at java.base/java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecutor.java:628)
	at java.base/java.lang.Thread.run(Thread.java:829)
Caused: org.jvnet.hudson.reactor.ReactorException
	at org.jvnet.hudson.reactor.Reactor.execute(Reactor.java:290)
	at jenkins.InitReactorRunner.run(InitReactorRunner.java:49)
	at jenkins.model.Jenkins.executeReactor(Jenkins.java:1205)
	at jenkins.model.Jenkins.<init>(Jenkins.java:992)
	at hudson.model.Hudson.<init>(Hudson.java:86)
	at hudson.model.Hudson.<init>(Hudson.java:82)
	at hudson.WebAppMain$3.run(WebAppMain.java:247)
Caused: hudson.util.HudsonFailedToLoad
	at hudson.WebAppMain$3.run(WebAppMain.java:264)
```