@Vlatombe 2023-08-31
Filed https://github.com/jenkinsci/ec2-plugin/pull/883 to fix the impacted plugin. Given this happened already with 2 plugins overriding `readResolve` without calling its super implementation, I will also file a robustness improvement for this change.