@Vlatombe 2023-08-31
Filed https://github.com/jenkinsci/jenkins/pull/8448